pyproject.toml
[project]
name = "kis-portfolio"
version = "0.1.0"
description = "한국투자증권 Open API 기반 포트폴리오 서비스"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "migusdn", email = "migusdn@gmail.com" }
]
dependencies = [
    "authlib>=1.6.1",
    "cryptography>=45.0.7",
    "duckdb>=1.5.2",
    "fastmcp>=2.5.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "mcp>=1.9.1",
    "pathlib>=1.0.1",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
kis-portfolio-auth = "kis_portfolio.adapters.auth:main"
kis-portfolio-remote = "kis_portfolio.remote:main"
kis-portfolio-mcp = "kis_portfolio.adapters.mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kis_portfolio"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
